Move definitions of `Fwd` and `Bwd` to `client/typing.py` (#2669)
