+++ src/py/flwr/client/typing.py
@@ -14,8 +14,30 @@
 # ==============================================================================
 """Custom types for Flower clients."""
 
+from dataclasses import dataclass
 from typing import Callable
+
+from flwr.client.workload_state import WorkloadState
+from flwr.proto.task_pb2 import TaskIns, TaskRes
 
 from .client import Client as Client
 
+
+@dataclass
+class Fwd:
+    """."""
+
+    task_ins: TaskIns
+    state: WorkloadState
+
+
+@dataclass
+class Bwd:
+    """."""
+
+    task_res: TaskRes
+    state: WorkloadState
+
+
+FlowerCallable = Callable[[Fwd], Bwd]
 ClientFn = Callable[[str], Client]
